[fix] fix date structure
